feat: rename rotation_matrix to rot_mat and add rotation matrix functions
